% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNLDAR.R
\name{myKNN}
\alias{myKNN}
\title{KNN method to predict test set using R}
\usage{
myKNN(train_x, train_y, test_x, k = 5)
}
\arguments{
\item{train_x}{the properties of the train set}

\item{train_y}{the labels of the train set}

\item{test_x}{the test set to be predicted}

\item{k}{the number of data used to predict test set}
}
\value{
a label for test set
}
\description{
KNN
}
\examples{
\dontrun{
data(data)
attach(data)
ypred <- myKNN(X, y,test_x = X, k = 5)
}
}
