% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNLDAR.R
\name{myLDA}
\alias{myLDA}
\title{LDA method to predict test set using R}
\usage{
myLDA(train_X, train_y, test_x)
}
\arguments{
\item{train_X}{the properties of the train set}

\item{train_y}{the lables of the train set}

\item{test_x}{the test set to be predicted}
}
\value{
a label for test set
}
\description{
LDA
}
\examples{
\dontrun{
data(data)
attach(data)
ypred <- myLDA(X, y,test_x = X)
}
}
